separated out backend code
