+++ backend/jest.config.ts
@@ -2,6 +2,7 @@ import type {Config} from 'jest';
 
 const config: Config = {
     preset: 'ts-jest',
+    testMatch: ["**/*.test.ts"]
 };
 
 export default config;
