+++ backend/src/__tests__/users.test.ts
@@ -1,4 +1,4 @@
-import { authController } from '../controllers/authController';
+import { userController } from '../controllers/userController';
 import pool from '../database/pool';
 
 jest.mock('../database/pool', () => ({
@@ -6,7 +6,7 @@ jest.mock('../database/pool', () => ({
 }));
 
 jest.mock('bcrypt', () => ({
-    compare: jest.fn((x, y) => x === y)
+    hash: jest.fn((x) => x)
 }));
 
 const mockedRequest = {
@@ -26,7 +26,7 @@ const mockedClient = {
     release: jest.fn()
 };
 
-describe('Tests for authentication api', () => {
+describe('Tests for user api', () => {
     beforeEach(() => {
         jest.clearAllMocks();
     })
@@ -38,7 +38,7 @@ describe('Tests for authentication api', () => {
             }
         };
 
-        await authController(testRequest, mockedResponse);
+        await userController(testRequest, mockedResponse);
         expect(mockedResponse.status).toHaveBeenCalledWith(400);
         expect(mockedResponse.json).toHaveBeenCalled();
     });
@@ -50,48 +50,31 @@ describe('Tests for authentication api', () => {
             }
         };
 
-        await authController(testRequest, mockedResponse);
+        await userController(testRequest, mockedResponse);
         expect(mockedResponse.status).toHaveBeenCalledWith(400);
         expect(mockedResponse.json).toHaveBeenCalled();
     });
 
-    it ('should send 400 status code when user does not exist', async () => {
+    it ('should send 400 status code when user already exists', async () => {
         const mockedRow = {
-            rows: []
+            rows: [mockedRequest.body]
         }
 
         mockedClient.query.mockResolvedValueOnce(mockedRow);
         
-        await authController(mockedRequest, mockedResponse);
-        expect(mockedResponse.status).toHaveBeenCalledWith(400);
-        expect(mockedResponse.json).toHaveBeenCalled();
-    });
-
-    it ('should send 400 status code when password is incorrect', async () => {
-        const user = {
-            username: 'test_username',
-            password: 'different_password'
-        }
-
-        const mockedRow = {
-            rows: [user]
-        }
-
-        mockedClient.query.mockResolvedValueOnce(mockedRow);
-
-        await authController(mockedRequest, mockedResponse);
+        await userController(mockedRequest, mockedResponse);
         expect(mockedResponse.status).toHaveBeenCalledWith(400);
         expect(mockedResponse.json).toHaveBeenCalled();
     });
 
     it ('should send 200 status code when credentials are correct', async () => {
         const mockedRow = {
-            rows: [mockedRequest.body]
+            rows: []
         }
 
         mockedClient.query.mockResolvedValueOnce(mockedRow);
 
-        await authController(mockedRequest, mockedResponse);
+        await userController(mockedRequest, mockedResponse);
         expect(mockedResponse.status).toHaveBeenCalledWith(200);
         expect(mockedResponse.json).toHaveBeenCalled();
     });
@@ -100,7 +83,7 @@ describe('Tests for authentication api', () => {
         const spy = jest.spyOn(pool, 'connect');
         spy.mockImplementationOnce(jest.fn(() => {throw Error()}));
         
-        await authController(mockedRequest, mockedResponse);
+        await userController(mockedRequest, mockedResponse);
         expect(mockedResponse.status).toHaveBeenCalledWith(500);
         expect(mockedResponse.json).toHaveBeenCalled();
     });
